fix warning about duplicate variables
When subsetting a data.table with column names in a character vector,
the `..` syntax needs to be used or data.table will think that the given
symbol is the column name. However that introduces a conflict between
the practice of setting column names to `NULL` to avoid R CMD check
warnings about using unset variables and the data.table practice of
expecting `..var` to refer to a variable `var` in the calling
environment. If `..var` is set to `NULL` in the calling environment and
`..var` is used in `[.data.table` then data.table will warn that both
`..var` and `var` are defined in the calling environment. The solution
is to use `with = FALSE` in `[.data.table` to avoid NSE.