remove duplicate samples when getting trio dCR
Sometimes samples can be parts of multiple batches (if different
clusterings were used) and that will cause the sample to appear in
multiple dCR matrices which causes issues with column name mangling if
those matrices are merged. So when we get the dCR matrices for a trio in
the de novo annotation we have to make sure samples aren't included
multiple times.