use union of dCR when analyzing de novo
Previously, when combining the dCR matrices for a child, father,
mother trio, only the intersection of their dCR bins was used to
regenotype. This could result in some inaccuracies when computing per
sample means because some bins could be omitted. Now we keep a union of
all bins and the dCR value at bins not available in a sample are `NA`.