move writing lines into WDL command section
Writing lines of an Array of Files will write the paths differently
based on where in the WDL it is called. If called in a non-inputs
declaration section, the paths will be written as remote paths. If
called in a command section, the paths will be written as local paths,
which is what script de novo calling script requires.