add df_to_gr method for data.table
The `df_to_gr` method for `data.frame`s assigns to the metadata slot of
the constructed `GRanges` object by creating a character vector of
column names, subsetting the `data.frame`, and then assigning the
resulting `data.frame`. However, due to the NSE mechanics of the
`[.data.table` method, `data.table` columns cannot be subset with a
character vector in the same way as a `data.frame`. Thus, we add a
separate `df_to_gr` method for `data.table`s to handle the special
subsetting method.