fix `data.table` to `GRanges` conversion
A `data.table` without metadata columns will cause an error during
conversion to a `GRanges` object because metadata columns are assigned
unconditionally. If there aren't any metadata columns, an empty
`data.table` will be created and assigned to the `GRanges` object and
cause an error about a zero-length object being assigned to a
nonzero-length object.