move line continuation into placeholder
Line continuation backslashes are used in the WDL command section after
each flag/option to the R script. The backslashes are added
unconditionally because they are outside the conditional placeholder
which results in a unnecessary backslashes if the corresponding option
is not given to the WDL. This commit places the backslashes inside the
placeholder to see if they will only be included if necessary.