Scale memory with number of dCR files
The memory requirements of the script to generate the ploidy matrix
scales with the number of samples and contigs. The number of contigs
will not increase beyond the default because the default includes all
human chromosomes so the main scaling factor is the number of samples.
The best estimate would be to count the number of samples in the
callset, but that would require setting up a task to do the counting. A
simpler estimate is to count the number of dCR matrices because the
number of matrices generally scales with the number of samples.