Change plot type for single interval dCR plots
Problem:
When the dCR matrices for families with members in different batches are
merged, it is possible that there is only a single common dCR interval
among the family members. In this case, the matplot function will not
plot any data because it needs at least two intervals and an empty plot
will be written.

Solution:
Plot families with only one common interval as a box-and-whisker plot.