fix plotting of single points in de novo plots
Problem:
The method to determine which intervals in the de novo trio dCR matrix
need to be plotted as points relies on finding isolated intervals: those
with a missing dCR value in the preceding and following interval. The
method to find those intervals fails in the case where there is only
one interval with a dCR value.

Solution:
Fix the method to find isolated intervals so that they can be correctly
plotted.