dynamically compute runtime memory
Problem:
The CheckContigsPloidy task uses a fixed amount of runtime memory which
does not scale with input size and would require overriding the runtime
attributes for larger input sizes.
Solution:
Scale runtime memory with the number of samples, which is the main
memory growth factor. The scaling factors were not chosen with any
thorough testing. The task completed with 4 GB at 4000 samples and
failed with 4 GB at ~5000 samples so 1 GB per 1000 samples plus padding
was chosen.