de novo: propogate errors from mclapply
Errors that occur in the child processes created by an mclapply call
will not be propogated and the conditions will just become part of the
return value of mclapply. It is better to explicily check the return
value for error conditions and fail if present so they are not masked.