annotate_denovo: fix filtering of high-quality calls
Subsetting the callset to high-quality calls using a single column did
not work because the name of the column by which to filter was stored in
a variable and the variable was not properly being evaluated in the
calling scope when subsetting the `data.table` callset. Using `eval`
gets around the issue.